modal confirmation delete post feed

+++ Client/src/components/Feed/PostListItem.jsx
@@ -16,13 +16,27 @@ import { Link } from "react-router-dom";
 import { TrashFill } from "react-bootstrap-icons";
 import "moment/locale/es";
 import moment from "moment";
+import Swal from "sweetalert2";
 
 export default function PostListItem(props) {
 	const { post } = props;
 	const { deletePost } = props;
 
 	const handleDelete = (id, image) => {
-		deletePost(id, image);
+		Swal.fire({
+			title: `¿Deseas borrar tu cuenta?`,
+			text: "Esta acción no se puede revertir.",
+			icon: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#3085d6",
+			cancelButtonColor: "#d33",
+			confirmButtonText: "Borrar",
+			cancelButtonText: "Cancelar",
+		}).then((result) => {
+			if (result.isConfirmed) {
+				deletePost(id, image);
+			}
+		});
 	};
 
 	return (
